+++ modules/math/sqrt/sqrt.c
@@ -3,7 +3,7 @@
 #include <math.h>
 
 r32 sqrt__r32(r32 r) {
-    return sqrt(r);
+    return sqrtf(r);
 }
 
 r64 sqrt__r64(r64 r) {
